made auto base to decrease repetition
