+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/autos/autoclip/AutoClipBase.java
@@ -6,11 +6,10 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import org.firstinspires.ftc.robotcore.external.navigation.Position;
 import org.firstinspires.ftc.robotcore.external.navigation.YawPitchRollAngles;
 import org.firstinspires.ftc.teamcode.Robot;
+import org.firstinspires.ftc.teamcode.opmodes.autos.AutoBase;
+import org.firstinspires.ftc.teamcode.opmodes.autos.autopark.AutoParkBase;
 
-public class AutoClipBase extends OpMode {
-
-    Robot robot = new Robot();
-
+public class AutoClipBase extends AutoBase {
     int stage = 1;
     boolean parking = false;
     final double TILE = 24;
@@ -40,8 +39,7 @@ public class AutoClipBase extends OpMode {
 
     @Override
     public void init() {
-        robot.init(hardwareMap);
-        robot.manipulatorArm.toggleManipulatorState();
+        super.init();
     }
 
     public void start() {
@@ -49,24 +47,6 @@ public class AutoClipBase extends OpMode {
     }
 
     public void loop() {
-        // USING GLOBAL TO BE IDEMPOTENT
-        if (!robot.globals.getManipArmAccurate()) {
-            robot.manipulatorArm.setTargetArmRotation(40);
-            if (robot.manipulatorArm.getCurrentArmRotation() >= 40) {
-                robot.manipulatorArm.rotationMotor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-                robot.manipulatorArm.setTargetArmRotation(0);
-                robot.globals.setManipArmAccurate(true);
-            }
-        }
-        // END -- THIS IS IDEMPOTENT and CAN BE RUN MORE THAN ONCE PER MATCH
-
-        robot.updatePosition();
-
-        Position position = robot.getPosition();
-        YawPitchRollAngles angles = robot.getOrientation();
-        telemetry.addData("Robot says Position", position.toString());
-        telemetry.addData("Robot says Orientation", angles.toString());
-        telemetry.addData("Camera says Live", robot.camera.isLive());
         if (getAgeInSeconds() < 25) {
             switch (stage) {
                 case 1:
@@ -121,13 +101,7 @@ public class AutoClipBase extends OpMode {
             parking = true;
         }
 
-        //TODO: DO DIS
-        //plans:
-        //wait a little bit to allow alliance to go
-        //go left until near basket zone
-        //turn to drop piece
-        //drop piece
-        //go to hang zone
+        super.loop();
     }
 
     public int getAgeInSeconds() {
