+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/autos/autopark/AutoParkBase.java
@@ -6,11 +6,9 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import org.firstinspires.ftc.robotcore.external.navigation.Position;
 import org.firstinspires.ftc.robotcore.external.navigation.YawPitchRollAngles;
 import org.firstinspires.ftc.teamcode.Robot;
+import org.firstinspires.ftc.teamcode.opmodes.autos.AutoBase;
 
-public class AutoParkBase extends OpMode { // TODO: TEST THIS
-
-    Robot robot = new Robot();
-
+public class AutoParkBase extends AutoBase { // TODO: TEST THIS
     final double TILE = 24;
 
     double PARK_X = 0.0; // CHANGE_ME!!
@@ -20,31 +18,13 @@ public class AutoParkBase extends OpMode { // TODO: TEST THIS
 
     @Override
     public void init() {
-        robot.init(hardwareMap);
-        robot.manipulatorArm.toggleManipulatorState();
+        super.init();
     }
 
     public void loop() {
-        // USING GLOBAL TO BE IDEMPOTENT
-        if (!robot.globals.getManipArmAccurate()) {
-            robot.manipulatorArm.setTargetArmRotation(40);
-            if (robot.manipulatorArm.getCurrentArmRotation() >= 40) {
-                robot.manipulatorArm.rotationMotor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-                robot.manipulatorArm.setTargetArmRotation(0);
-                robot.globals.setManipArmAccurate(true);
-            }
-        }
-        // END -- THIS IS IDEMPOTENT and CAN BE RUN MORE THAN ONCE PER MATCH
-
-        robot.updatePosition();
-
         if (!ACHIEVED_POSITION) {
-            Position position = robot.getPosition();
-            YawPitchRollAngles angles = robot.getOrientation();
-            telemetry.addData("Robot says Position", position.toString());
-            telemetry.addData("Robot says Orientation", angles.toString());
-            telemetry.addData("Camera says Live", robot.camera.isLive());
             ACHIEVED_POSITION = robot.mecanumDrive.driveToPosition(PARK_X, PARK_Y, position, angles);
         }
+        super.loop();
     }
 }
